Some changes to allow for building on mac

+++ Engine/src/Core/Settings.cpp
@@ -162,9 +162,9 @@ int Settings::GetInt(const char* section, const char* key)
 	if (s_DefaultValues.find({ section, key }) == s_DefaultValues.end())
 	{
 		ENGINE_WARN("No default setting for [{0}] {1}", section, key);
-		return s_Ini->GetLongValue(section, key);
+		return (int)s_Ini->GetLongValue(section, key);
 	}
-	return s_Ini->GetLongValue(section, key, atoi(s_DefaultValues.at({ section, key }).c_str()));
+	return (int)s_Ini->GetLongValue(section, key, atoi(s_DefaultValues.at({ section, key }).c_str()));
 }
 
 /* ------------------------------------------------------------------------------------------------------------------ */
